+++ src/app/sections/map/map-box/map-box.component.html
@@ -1,8 +1,6 @@
-<div class="map" id="map">
-  <!-- <div id="map-container">
-    <div id="top-bar">
-      <p>Prueba de top bar</p>
-    </div>
-    <div id="bottom-bar"></div>
-  </div> -->
+<div class="map" id="map"></div>
+<div id="map-container">
+  <div id="bottom-bar">
+    <p>Prueba de top bar</p>
+  </div>
 </div>
