+++ src/app/sections/map/map-box/map-box.component.sass
@@ -8,24 +8,19 @@
 
 #map-container
     position: absolute
-    top: 0
     bottom: 0
-    width: 100%
-    z-index: 0
-    white-space: normal
-
-#top-bar
-    position: relative
+    right: 0
     z-index: 2
+    background: rgba(255, 255, 255, 0.4)
+    font-family: Arial, sans-serif
+    overflow: auto
+    border-radius: 3px
+    width: 100%
 
 #bottom-bar
-    position: relative
+    top: 0
+    height: 80px
+    margin-top: 20px
+    width: 100%
     z-index: 2
-
-.marker 
-  background-image: url('https://cdn0.iconfinder.com/data/icons/small-n-flat/24/678111-map-marker-512.png');
-  background-size: cover;
-  width: 50px;
-  height: 50px;
-  border-radius: 50%;
-  cursor: pointer;
+
