--- src/app/sections/map/map.spec.ts
@@ -1,7 +1,0 @@
-import { Map } from './map';
-
-describe('Map', () => {
-  it('should create an instance', () => {
-    expect(new Map()).toBeTruthy();
-  });
-});
