+++ src/app/services/map.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { environment } from '../../environments/environment';
 
-import { GeoJson } from '../sections/map/map';
+import { GeoJson } from '../shared/models/Map';
 import * as mapboxgl from 'mapbox-gl';
 
 @Injectable()
