+++ src/app/shared/models/Map.ts
